Fix: scheduler: Honor kind=Mandatory for symmetrical constraints
Previously, mandatory symmetrical stop constraints were treated as
optional. It seems likely that **all** mandatory constraints were
treated as optional, but the issue was masked for start and promote
actions. Constraints for these actions seemed to be honored as
mandatory due to a runnability dependency: the right resource couldn't
start if the left resource was unrunnable.

This patch fixes the issue by reversing the `implies` directions in
`get_flags()`.

Asymmetrical constraints still are not enforced as mandatory.