cluster: try to reconnect with corosync if it goes down
If the corosync cfg conection goes down, don't just quit, but
keep trying to reconnect.

In theory this could retry forever, but in practice the attrd->cib
connection times out. I've increased the timeout for this to give
corosync a little longer to start, but it's that timeout that
ultimately decides how long pacemaker will survive the loss of
corosync.

(In actual practice, of course, the node should be fenced!)