Fix: libpacemaker: Don't shuffle anon clones when choosing a node
Currently, anonymous clone instances may be shuffled when a clone is
newly starting on a node with a higher allocation score for that
resource.

For example, this can happen for a promotable clone when the resource is
starting on a node with a higher promotion score. Let's say instance 0
is running on node2 in non-promoted state, node1 has a higher promotion
score, and the clone is stopped on node1 and about to start there. Then
instance 0 gets allocated to node 1 and moved there, and instance 1
starts on node2. This shuffling causes a restart on node2.

The fix is to favor the current node for an anonymous clone instance
even if there's a node with a higher weight available. It doesn't matter
which particular instance runs on any one node, so it's better to leave
an instance where it's currently running.

Now, in the example scenario described above, instance 0 remains on
node2, and instance 1 gets allocated to node1, where it will start and
promote.

Also update regression tests accordingly.
  - clone-anon-recover-promoted captures the main intended change.
  - 1360 and bug-cl-5168 appeared to be incorrect before the change.

Resolves: RHBZ#1931023

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>