Refactor: Controller: Drop redundant execd state call when client is DC
The DC also acts as a client (to itself) during the join sequence.
Currently, in do_cl_join_finalize_respond(), we get the local executor
history on the DC; if this fetch fails, we enter an error state and
don't send the client join confirmation.

Then in do_dc_join_ack(), we get the local executor history again and
use the first one (guaranteed non-NULL at this point) only as a
fallback. This was a fix via 8032f79 for CLBZ#5286 (which incidentally
looks like quite similar to RHBZ#2010084, but the fix affects only the
DC).

Since the first and second fetches use the same function call on the
same node, we can reasonably expect that either both will succeed or
both will fail most of the time. The only apparent advantage of the
first fetch is that if it fails, we avoid the overhead of one message
and of a crm_get_peer() call in do_dc_join_ack().

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>