Refactor: based: Allow queueing of cluster messages
Allow pacemaker-based to lock itself with respect to messages received
from the cluster. While locked, pacemaker-based queues cluster messages
to process later when an unlock request is received.

If any messages remain in the queue when pacemaker-based is shut down,
they are destroyed without being processed.

This functionality is cluster stack-agnostic, even though only corosync
is currently supported. The idea is to lock against messages received
from other nodes.

If needed, we can later add the ability to lock the CIB with respect to
IPC messages and all messages. The same queue should be used for all
types of messages. Requests should be processed in the order in which
they were received.

If we add support for locking IPC messages, then we must ensure that we
can still process "unlock" messages while IPC messages are locked. We
would also need to ensure that one daemon does not unlock the CIB when
another daemon has locked it. Some options:
  * Do this with care at the caller side since it's all internal.
  * Require the client that requested the lock to unlock it. But this is
    dangerous: if we somehow lose the client, we can't unlock.
  * Implement IPC message locking per daemon instead of for IPC messages
    as a group. This seems like the best option. If we do this, that
    raises the question of whether to allow locking cluster messages
    per-daemon as well.

TODO: What happens if a daemon is killed (especially hard-killed) after
locking the CIB? We can unlock for the daemon at daemon init time, but
we can't safely unlock for the cluster or for other daemons at daemon
init time.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>