Refactor: libpacemaker: Get colocation attr from container host
If a node is a bundle node and the resource we're updating allowed nodes
for is a bundled resource, look up the colocation attribute on the
bundle node's assigned host. Likewise if the node is a guest node and
the resource has container-attribute-target set to "host".

For guest nodes, this basically aligns with the existing specification
of container-attribute-target.

Only the bundled resource can run on a bundle node. So if the bundled
resource is involved in a colocation (for example, "rscB with promoted
bundled-rscA"), then the colocation must be considered for the bundle
node's assigned host.

Ref T672

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>