Doc: daemons: Add missing periods between longdesc and allowed values
Usually longdescs end in a period already if given explicitly. However,
if a longdesc is NULL, the shortdesc is used as the longdesc. Shortdescs
usually don't end in a period.

If an option has allowed values, they're appended after the longdesc.
This leads to a situation like

"<longdesc> ... does not have quorum  Allowed values: ... </longdesc>"

in the meta-data and man pages.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>