Refactor: libpe_status: Drop dead "#default" code
The "#default" handling code was added by commit ea1359b. It seems that
it was never documented except as an allowed value for the target-role
meta-attribute. In that use case, it's now considered an invalid value.

The documentation for "target-role=#default" was only in the crm.dtd
file (not in the doc books), which was removed in 2.0.0 by commit
79a739c7.

Any user relying on it is relying on undefined and undocumented
behavior.

We also have to update some tests. When we drop the "#default" line from
add_hash_param, a "#default" value gets added to a resource's
meta-attributes if present in the XML. This results in a resource
definition change, which would change two test outputs.

If we want to be especially generous, we can wait for a compatibility
break. I feel like this has been undocumented and useless for long
enough.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>