Refactor: controller: Drop dead CRM_OP_LOCAL_SHUTDOWN
The only thing that sent CRM_OP_LOCAL_SHUTDOWN was crmadmin, and that
was dropped by commit 5919445. It was only ever local, so we don't have
to worry about getting the op from another node.

Also make crm_shutdown static and drop the forward declaration, since
it's only used in one file now.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>