Refactor: tools: Simplify is_scheduler_required() in crm_resource
The set of commands in the default case contains the entire set of
commands for which get_find_flags() returns a nonzero value. So the
((get_find_flags != 0) && (options.rsc_id != NULL)) condition is
redundant.

The one exception was cmd_set_property. However, that case could never
have been hit in the switch statement. cmd_set_property has find flags
associated with it and requires a --resource argument (so options.rsc_id
!= NULL).

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>