Fix: libcrmcommon: Escape special chars in XML attribute text output
Second half of quasi-regression introduced by c6d9cea.

When displaying XML attributes in text output format, escape special
characters in the attribute value. (An attribute name isn't allowed to
have special characters.) This ensures the entire attribute value stays
on one line, and that it doesn't have (for example) a double quote
inside the attribute value.

attrd_updater and crm_attribute use this output method. It turns out
that the openstack-info resource agent sets a multiline attribute (which
I consider a bug, but it's longstanding behavior). The
openstack-virtual-ip resource agent then depends on being able to parse
it from the attrd_updater text output as **one line**.

If there's a literal newline character in the displayed value, then it
of course gets split over two lines, breaking the agent's parsing.

(The openstack-info agent should set the attribute to a saner value, and
the openstack-virtual-ip agent should parse the XML output instead of
the text output, but here we are.)

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>