Refactor: libcib: Drop dedupXpathResults() internally
This addresses what ca271b67 addressed, but without the need to
deduplicate results. I took this idea from xpath2.c:update_xpath_nodes()
in libxml2's examples.

If cib_multiple is not set, we continue to iterate over the results in
normal (forward) order. Existing callers expect us to use the first
result in that case.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>