Refactor: tools: Improve error handling in crm_diff.c
We could use g_set_error() for the "Could not apply patch" error, but
that feels like a slippery slope. That's basically saying there was an
error in libcrmcommon or (in the future perhaps libpacemaker).

Typically we call out->err() directly in those cases (from within the
library once there's a function that accepts a pcmk__output_t object),
and we don't bother reporting the error at the CLI tool level.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>