Refactor: libcrmcommon: New pcmk__xml_commit_changes()
To replace xml_accept_changes(). For now we basically mimic the logic in
xml_accept_changes(), but perhaps we should be less aggressive or rename
the function to reflect that it's clearing so many things.

In particular, I'm not sure I like that this disables tracking, and I
don't think I like that ACLs are so tightly coupled to change tracking
(in many ways, not just here). Change tracking is required for ACLs but
not vice-versa. Perhaps operations and flags should be kept a bit
more... separate, though.

Still, it's probably wise to throw away any unpacked ACLs when
committing changes, as we're doing now and as we've been doing. ACLs
themselves may have been added, modified, or removed as part of the
changes. We want to get a current view of them before any further
changes.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>