Refactor: libcrmcommon: Consolidate both-non-NULL in mark_xml_changes()
Move the old_child != NULL code in the new_child loop, up to the
old_child loop. This is another case where old_child and new_child are
both non-NULL; we already cover that case in the first loop.

There is a corner case here, where behavior changes but neither the old
behavior nor the new behavior seems correct. match_xml() doesn't
actually require an exact match, and its matching behavior is
asymmetric. This will be addressed in an upcoming commit.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>