Refactor: various: Check pcmk__xe_get_time() rc only where useful
Take advantage of the fact that the destination output variable is
unchanged on error. If the caller initialized it to some value (either
explicitly or by allocating memory with calloc() or
pcmk__assert_alloc()), then we can assume that either it's unchanged or
a valid value was parsed into it.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>