Refactor: libcrmcommon: Allow negative values in pcmk__parse_ms()
All callers except crm_mon already handled this.

Also, crm_mon was the only caller that passed NULL for the dest
argument. So we rename it to result, get rid of the local result
variable, and CRM_CHECK that it's not NULL. It's easy to add "NULL dest"
as a feature again in the future, if there's a frequent need to validate
that a time+units string is parsable without actually needing its value.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>