Feature: libcrmcommon: Allow configuring unlimited fencing-max-attempts
...cluster property.

Prior to this commit, the only valid values were INFINITY, +INFINITY,
and strictly positive integers. Any other value would be rejected by the
validator function and would result in the default being used.

Now, nonpositive values (0, a negative integer, or -INFINITY) result in
unlimited fencing attempts.

This is mostly for convenience. The fencing-max-attempts property has
been documented as a score type for some time, because INFINITY and
+INFINITY are supported. We should have a way to configure unlimited
fencing attempts -- INFINITY is actually 1000000, but that's effectively
unlimited. But semantically, this is not a score.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>