Doc: libcrmcommon: Clarify chown_logfile() comments
For whatever reason, the intent here is to ensure that the daemon group
owns the log file or can read and write it.

There are other ways that the daemon user could read and write the file.
For example, the daemon user could own the file even if a different
group owns it, or the group owner could be one of the daemon user's
secondary groups. It's not clear that we should be changing the
ownership in those cases. Also, the daemon group could have read/write
permissions via ACLs.

The current logic ensures that ANY member of the daemon group can read
and write the file, even if it is a bit coarse in how it does so.

If we ever decide that we only want to ensure that the daemon user can
access the file, then we could impersonate the daemon user, try to open
the file, and then switch back to root.

There are a couple of other points in the "\todo" directive in Doxygen.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>