Refactor: libcrmcommon: Simplify pcmk__add_logfile()
Don't bother opening in append mode with fopen() unless we're root. In
that case, do it in set_logfile_permissions() to limit scope of the file
descriptor. We use that file descriptor only for passing to fstat(),
fchown(), and fchmod().

This also lets us simplify the error message, since the user must be
root, and lets us drop a fileno() error check.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>